refactor all floating viz stuff out of MapAnalysis
